factory
demonstrates the Factory pattern, which provides an interface for creating objects in a superclass, while allowing subclasses to alter the type of objects that will be created. In this example, Product is the superclass of ConcreteProductA and ConcreteProductB, which represent different types of products. Creator is the superclass of ConcreteCreatorA and ConcreteCreatorB, which create instances of the products. The factory_method method is overridden by the subclasses to specify which product to create.