iterator
demonstrates the Iterator pattern, which provides a way to access the elements of an aggregate object sequentially without exposing its underlying representation. In this example, Iterator is the iterator that provides the next element of the collection each time it is called. Collection is the aggregate object that holds a collection of items and provides an iterator for it